CORE-7550 Rename permIdRequest.model.PermanentIdRequestPathProvider
Renamed permIdRequest.views.FolderPathProvider to
permIdRequest.model.PermanentIdRequestPathProvider.
